+++ x/ratesync/keeper/ibc_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/cosmos/gogoproto/proto"
 	icatypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/types"
 	channeltypes "github.com/cosmos/ibc-go/v7/modules/core/04-channel/types"
+
 	"github.com/persistenceOne/pstake-native/v2/x/ratesync/keeper"
 	"github.com/persistenceOne/pstake-native/v2/x/ratesync/types"
 )